// src/types/response.ts
export type Gender = 'F' | 'M' | 'NB' | 'NL' | 'UN';
export type Role = 'DEV' | 'UX' | 'LEAD';
export type RoleExperience = '<1' | '1-3' | '3-5' | '>5';
export type AgileExperience = 'NEVER' | 'OCCASIONAL' | 'USUAL';
export type ProjectType = 'WEB' | 'MOBILE' | 'OTHER';
export type SprintPlanningExperience =
  | 'USUAL'
  | 'SOMETIMES'
  | 'ONLY_CRITICAL'
  | 'NEVER';